Update to use funcx-common pubsub wrapper
Replace the RedisPubSub object defined in this repo with the wrapper
for pubsub provided by funcx-common. The API it provides is nearly
identical, but there is no explicit `connect()` call -- the
connection to redis is handled during init.

Move the 'RedisTask' definiiton module up several layers, as it is now
the only thing in the 'queues/redis/' dir. That structure doesn't make
much sense with the current content.