Make changes to get flake8 linting to pass
flake8 updates are mostly line-breaking, often in comments and
multiline docstrings.

In some cases, while reformatting log statements, it's also expedient
to convert them to use the logging %-formatting, which defers
rendering until the log record is emitted (and therefore is more
efficient for debug-level logs). This is done in cases in which
the AST would need to be modified non-trivially anyway.